[ET-VK] Move `ParamsBuffer` and `StorageBuffer` to standalone files
Includes the renaming of `UniformParamsBuffer` to `ParamsBuffer` for brevity.

These objects aren't tightly coupled to `Context` and hence they are better placed in standalone files.

Differential Revision: [D59281543](https://our.internmc.facebook.com/intern/diff/D59281543/)

[ghstack-poisoned]
