+++ backends/vulkan/runtime/graph/ComputeGraph.h
@@ -95,7 +95,7 @@ class ComputeGraph final {
   std::unique_ptr<api::Context> context_;
   std::vector<SharedObject> shared_objects_;
   std::vector<Value> values_;
-  std::vector<api::UniformParamsBuffer> param_ubos_;
+  std::vector<api::ParamsBuffer> param_ubos_;
 
   std::vector<std::unique_ptr<PrepackNode>> prepack_nodes_;
   std::vector<std::unique_ptr<ExecuteNode>> execute_nodes_;
@@ -382,7 +382,7 @@ class ComputeGraph final {
 
   template <typename Block>
   const api::BufferBindInfo create_params_buffer(const Block& data) {
-    param_ubos_.emplace_back(api::UniformParamsBuffer(context_.get(), data));
+    param_ubos_.emplace_back(api::ParamsBuffer(context_.get(), data));
     return api::BufferBindInfo(param_ubos_.back().buffer());
   }
